backend/api-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
sqlmodel==0.0.22
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
alembic==1.13.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.11.7
pydantic-settings==2.5.2
google-generativeai==0.8.3
openai>=1.107.1,<2.0.0
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2025.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
